backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
pdfplumber==0.10.3
pydantic==2.5.3
reportlab==4.0.7
Pillow==10.1.0
google-generativeai==0.3.2
python-dateutil==2.8.2
pytz==2023.3
